zoombrain/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and indexing
faiss-cpu==1.7.4
numpy==1.24.3

# Document parsing
PyPDF2==3.0.1
python-pptx==0.6.23

# Machine learning / embeddings
sentence-transformers==2.2.2

# LLM support (optional - uncomment if using OpenAI)
# openai==1.3.0

# Email support
# secure-smtplib is built-in with Python

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0

# HTTP client for demo
requests==2.31.0

# Date parsing (optional enhancement)
# python-dateutil==2.8.2

# Google Calendar API (optional - uncomment if using)
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.108.0

# Slack integration (optional - uncomment if using)
# slack-sdk==3.23.0

